Git-Agent/agent_new/requirements.txt
streamlit==1.35.0
langchain==0.1.13
langchain-openai==0.1.1
openai==1.13.3
gitpython==3.1.43
PyGithub==2.2.0
python-dotenv==1.0.1
pytest==8.0.0